requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
asyncio==4.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.14.0
dacite==1.9.2
docopt-ng==0.9.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
mutagen==1.47.0
propcache==0.4.1
pyaes==1.6.1
pycparser==3.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.6
requests==2.32.5
scdl==3.0.4
SClib==1.0.0
six==1.17.0
soundcloud-lib==0.6.1
soundcloud-v2==1.6.2
soundcloud_downloader==0.1.4
soupsieve==2.8.3
TgCrypto==1.2.5
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
yt-dlp==2026.3.3